Ajout des pages pour ajouter et modifier des rapports, ainsi que des colonnes de tableau pour la gestion des personnes.
