+++ app/persons/[personId]/add-report/page.tsx
@@ -1,0 +1,13 @@
+export default async function PersonIdAddReportPage() {
+  return (
+    <div className="h-full">
+      <main>
+        <div className="mx-auto my-10 flex max-w-4xl items-center justify-between p-4">
+          <h1 className="text-center font-lato text-4xl font-black">
+            Ajoouter un rapport
+          </h1>
+        </div>
+      </main>
+    </div>
+  );
+}
