+++ src/components/persons/table.tsx
@@ -1,9 +1,8 @@
 "use client";
 
+import { personsTableColumns } from "@/components/persons/table-columns";
 import { PersonsTableFilter } from "@/components/persons/table-filter";
 import { PersonsTablePagination } from "@/components/persons/table-pagination";
-import { Button } from "@/components/ui/button";
-import { Checkbox } from "@/components/ui/checkbox";
 import {
   Table,
   TableBody,
@@ -12,10 +11,8 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
-import { formatPhoneNumber } from "@/lib/utils";
 import { Person } from "@prisma/client";
 import {
-  ColumnDef,
   ColumnFiltersState,
   flexRender,
   getCoreRowModel,
@@ -26,126 +23,7 @@ import {
   useReactTable,
   VisibilityState,
 } from "@tanstack/react-table";
-import { formatDistanceToNowStrict } from "date-fns";
-import { fr } from "date-fns/locale";
-import { ArrowUpDown, MoreHorizontalIcon } from "lucide-react";
 import { useState } from "react";
-
-export const personsTableColumns: ColumnDef<Person>[] = [
-  {
-    id: "select",
-    header: ({ table }) => (
-      <Checkbox
-        checked={
-          table.getIsAllPageRowsSelected() ||
-          (table.getIsSomeRowsSelected() && "indeterminate")
-        }
-        onCheckedChange={(value) => table.toggleAllPageRowsSelected(!!value)}
-        aria-label="Select all rows"
-        className="mr-2"
-      />
-    ),
-    cell: ({ row }) => (
-      <Checkbox
-        checked={row.getIsSelected()}
-        onCheckedChange={(value) => row.toggleSelected(!!value)}
-        aria-label="Select this row"
-      />
-    ),
-    enableSorting: false,
-    enableHiding: false,
-  },
-  {
-    id: "Nom",
-    accessorKey: "fullname",
-    header: ({ column }) => (
-      <div className="flex items-center justify-start">
-        <span className="font-lato text-sm font-semibold">Nom</span>
-        <Button
-          className="ml-0"
-          variant={"ghost"}
-          size={"icon"}
-          onClick={() => {
-            column.toggleSorting(column.getIsSorted() === "asc");
-          }}
-        >
-          <ArrowUpDown />
-        </Button>
-      </div>
-    ),
-    cell: (row) => (
-      <div className="text-left font-epilogue text-sm">
-        {row.getValue() as string}
-      </div>
-    ),
-  },
-  {
-    id: "E-mail",
-    accessorKey: "email",
-    header: () => <div className="font-lato text-sm font-semibold">E-mail</div>,
-    cell: (row) => (
-      <div className="truncate text-left font-epilogue text-sm">
-        {row.getValue() as string}
-      </div>
-    ),
-  },
-  {
-    id: "Téléphone",
-    accessorKey: "phone",
-    header: () => (
-      <div className="font-lato text-sm font-semibold">Téléphone</div>
-    ),
-    cell: (row) => {
-      return (
-        <div className="text-left font-epilogue text-sm">
-          {formatPhoneNumber(row.getValue() as string)}
-        </div>
-      );
-    },
-  },
-  {
-    id: "Statut",
-    accessorKey: "status",
-    header: () => <div className="font-lato text-sm font-semibold">Statut</div>,
-    cell: (row) => (
-      <div className="truncate text-left font-epilogue text-sm">
-        {row.getValue() as string}
-      </div>
-    ),
-  },
-  {
-    id: "Age",
-    header: () => <div className="font-lato text-sm font-semibold">Age</div>,
-    cell: ({ row }) => {
-      const person = row.original;
-      const age = formatDistanceToNowStrict(person.dateOfBirth, {
-        locale: fr,
-      });
-      // const age = 20;
-
-      return <div className="font-epilogue text-sm">{age}</div>;
-    },
-  },
-  {
-    id: "action",
-    header: () => (
-      <div className="mr-2 text-right font-lato text-sm font-semibold">
-        Actions
-      </div>
-    ),
-    cell: () => (
-      <Button
-        variant={"ghost"}
-        className="flex w-full justify-end p-0"
-        size={"icon"}
-      >
-        <MoreHorizontalIcon className="mr-2 size-4" />
-      </Button>
-    ),
-    enableSorting: false,
-    enableHiding: false,
-  },
-];
 
 export type PersonsTableProps = {
   persons: Person[];
@@ -222,7 +100,7 @@ export const PersonsTable = ({ persons }: PersonsTableProps) => {
                   colSpan={personsTableColumns.length}
                   className="h-24 text-center"
                 >
-                  No results.
+                  Aucun résultat.
                 </TableCell>
               </TableRow>
             )}
